[Ionut] - More bug fixes (mostly the ones resulting from switching to smart pointers)
git-svn-id: https://xp-dev.com/svn/Divide-Framework/trunk@360 e89c6373-eeba-4702-abef-d26284fe765c

Former-commit-id: 9fc6bbb990aa10bcd22e14dfe9514f88071f11d1 [formerly 2d5018f7287fe516fee960a4961e19847bd50d66] [formerly f6cee95c4d79ad84dbce05df2244e01bee02795f [formerly e9f17b52d973452329ad52896dcf94e64ade350a]]
Former-commit-id: 2c94fd4c246c74f029ae1945150a466041fc8423 [formerly 3286c8f46ebb35258f47fcdf1b3ac2ae7376ad5d]
Former-commit-id: 55b545323162fb28223b61bc4058131255f76bdc

+++ Source Code/Graphs/SceneGraph.cpp
@@ -6,9 +6,23 @@
 
 namespace Divide {
 SceneGraph::SceneGraph()
-    : _root(MemoryManager_NEW SceneGraphNode(MemoryManager_NEW SceneRoot(),
-                                             "ROOT"))
+    : _root(nullptr)
 {
+    load();
+}
+
+SceneGraph::~SceneGraph()
+{
+    unload();
+}
+
+void SceneGraph::load() {
+    if (_root != nullptr) {
+        return;
+    }
+
+    _root.reset(MemoryManager_NEW SceneGraphNode(MemoryManager_NEW SceneRoot(),
+                "ROOT"));
     _root->getComponent<RenderingComponent>()->castsShadows(false);
     _root->getComponent<RenderingComponent>()->receivesShadows(false);
     _root->setBBExclusionMask(TYPE_SKY | TYPE_LIGHT | TYPE_TRIGGER |
@@ -16,14 +30,17 @@ SceneGraph::SceneGraph()
                               TYPE_VEGETATION_TREES);
 }
 
-SceneGraph::~SceneGraph()
-{
+void SceneGraph::unload() {
+    if (_root == nullptr) {
+        return;
+    }
     Console::d_printfn(Locale::get("DELETE_SCENEGRAPH"));
     SceneNode* root = _root->getNode<SceneRoot>();
-    // Should recursively call unload on the entire scene graph
-    _root->unload();
     // Delete the root scene node
     MemoryManager::DELETE(root);
+    // Should recursively call unload on the entire scene graph
+    _root->unload();
+    _root.reset(nullptr);
 }
 
 void SceneGraph::idle() {
